[BOOKINGSG-9342][GZ] improvement - rename slot content dir, change children render
