+++ src/schedule/schedule-slot-content/hooks.ts
