+++ src/schedule/schedule-slot-content/index.ts
@@ -1,4 +1,4 @@
 export * from "./hooks";
 export * from "./schedule-slot-content";
-export * from "./time-slot-style";
+export * from "./style-helper";
 export * from "./utils";
