+++ src/schedule/schedule-slot-content/style-helper.ts
