+++ src/schedule/schedule-slot-content/utils.ts
