+++ src/schedule/schedule-slot-content/with-optional-popover.tsx
